refactor(reRefactor): Re refactoring logic
